% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{basic_plot}
\alias{basic_plot}
\title{Basic Plot + ribbon}
\usage{
basic_plot(
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  plot_group = NULL,
  color,
  title,
  xlab,
  ylab,
  add = NULL,
  CI_lab = "95\% CI"
)
}
\arguments{
\item{x}{x variable to plot}

\item{y}{y variable to plot}

\item{ymin}{optional minimum y value of ribbon}

\item{ymax}{optional maximum y value of ribbon}

\item{plot_group}{the plot group of the plot to belong to}

\item{color}{the plot and ribbon color}

\item{title}{title of plot}

\item{xlab}{the x axis label of the plot}

\item{ylab}{the y axis label of the plot}

\item{add}{plot to add on to (the object)}
}
\description{
Basic Plot + ribbon
}
