% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{bin_indices}
\alias{bin_indices}
\title{Bin indices}
\usage{
bin_indices(df, len.col, cols_to_bin, grouping, breaks, ...)
}
\arguments{
\item{df}{the data.frame to bin}

\item{len.col}{the length column to bin}

\item{cols_to_bin}{columns to summarize}

\item{grouping}{extra columns to group by}

\item{breaks}{the breaks of the binning}
}
\description{
Bin indices
}
