% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{compute_group_normalization}
\alias{compute_group_normalization}
\title{Normalize data to lie between 0 and 1}
\usage{
compute_group_normalization(
  data,
  group,
  var_y,
  var_x,
  normalizer = function(x) {
     (x - min(x))/(max(x) - min(x))
 },
  ycol = NULL
)
}
\arguments{
\item{data}{the data frame to operate on}

\item{group}{optional column to group data on when normalizing}

\item{var_y}{the variable to become between 0 and 1}

\item{var_x}{the x variable}

\item{normalizer}{the function to do the normalizing}

\item{ycol}{when not grouping, an extra column to include}
}
\description{
Used for plotly_ridges
}
