% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{create_report}
\alias{create_report}
\title{Create Report}
\usage{
create_report(
  title,
  model_file,
  report_output,
  tmb_data,
  mod_data,
  build = TRUE
)
}
\arguments{
\item{title}{the title of the report and file name of the report}

\item{model_file}{the path to the rds output of the model}

\item{report_output}{where to output the report}

\item{tmb_data}{path of tmb input data}

\item{mod_data}{path of modified data}

\item{build}{build the html file?}
}
\description{
Create a report from a model output file. Requires working Rmarkdown setup.
}
