% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{extract_from_sdr}
\alias{extract_from_sdr}
\title{Extracts specified variable name estimate and std. error from an sdreport summary}
\usage{
extract_from_sdr(sdr, var.name, ..., flatten = FALSE)
}
\arguments{
\item{sdr}{The sdreport summary to pull from}

\item{var.name}{The variable to extract}

\item{...}{Can specify dimensions as vectors, any number. MUST FOLLOW TEMPLATE ORDER & DIMENSIONS}

\item{flatten}{Convert array to flat data.frame}
}
\description{
Extracts specified variable name estimate and std. error from an sdreport summary
}
