% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{generate_CI}
\alias{generate_CI}
\title{Create  confidence interval}
\usage{
generate_CI(est, std, alpha = 0.05, undotrans = NULL)
}
\arguments{
\item{est}{thing of estimates}

\item{std}{thing of standard deviations}

\item{alpha}{value of alpha}

\item{undotrans}{function to undo transformation (e.g., exp)}
}
\description{
Create  confidence interval
}
