% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{get_K}
\alias{get_K}
\title{Convert from schnute VB k to traditional VB K}
\usage{
get_K(k)
}
\arguments{
\item{k}{the k parameter value of the Schnute VB growth curve}
}
\description{
Convert from schnute VB k to traditional VB K
}
