% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{get_L_infinity}
\alias{get_L_infinity}
\title{Get L_infty from Schnute VB}
\usage{
get_L_infinity(L, ell, k, M)
}
\arguments{
\item{L}{the largest length at age mean}

\item{ell}{the smallest length at age mean}

\item{k}{the Schnute k parameter}

\item{M}{the number of age classes}
}
\description{
Get L_infty from Schnute VB
}
