% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{get_mohns}
\alias{get_mohns}
\title{Calculate Mohn's rho and matrix for a specified quantity}
\usage{
get_mohns(peels.object, quantity, details = FALSE, get_years = TRUE, ...)
}
\arguments{
\item{peels.object}{the object containing all the peels}

\item{quantity}{name of quantity to get}

\item{details}{report the details?}

\item{get_years}{grab the years to use?}

\item{...}{things to pass to extract_from_sdr, if using years call it year}
}
\description{
Calculate Mohn's rho and matrix for a specified quantity
}
