% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{get_t_zero}
\alias{get_t_zero}
\title{Get t zero from Schnute VB}
\usage{
get_t_zero(L, ell, k, M, a1)
}
\arguments{
\item{L}{the largest length at age mean}

\item{ell}{the smallest length at age mean}

\item{k}{the Schnute k parameter}

\item{M}{the number of age classes}

\item{a1}{first age class in model}
}
\description{
Get t zero from Schnute VB
}
