% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{glogit}
\alias{glogit}
\title{Logistic function with different maximum}
\usage{
glogit(x, L)
}
\arguments{
\item{x}{the x parameter}

\item{L}{the maximum of the curve}
}
\description{
Logistic function with different maximum
}
