% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{iglogit}
\alias{iglogit}
\title{Inverse logistic function with different maximum}
\usage{
iglogit(y, L)
}
\arguments{
\item{y}{the thing to inverse}

\item{L}{the maximum of the curve}
}
\description{
Inverse logistic function with different maximum
}
