% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{inv_simplex_transform}
\alias{inv_simplex_transform}
\title{Inverse simplex transform with any maximum}
\usage{
inv_simplex_transform(y, L)
}
\arguments{
\item{L}{the maximum to hit}

\item{the}{vector to inverse}
}
\description{
Inverse simplex transform with any maximum
}
