% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{inv_unit_simplex_transform}
\alias{inv_unit_simplex_transform}
\title{Convert back to a vector that sums to 1}
\usage{
inv_unit_simplex_transform(y)
}
\arguments{
\item{y}{vector to convert back}
}
\description{
Convert back to a vector that sums to 1
}
