% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{make_corr_mat}
\alias{make_corr_mat}
\title{Create covariance matrix}
\usage{
make_corr_mat(cov.fixed, names = NULL)
}
\arguments{
\item{cov.fixed}{the covariance matrix (e.g from sd.report cov.fixed)}

\item{names}{optional vector of names to use}
}
\description{
Create covariance matrix
}
