% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{match_normalizer}
\alias{match_normalizer}
\title{Normalizer function to put two different data sets on the same scale in a ridgeline plot.}
\usage{
match_normalizer(survey1, survey2, year = NULL)
}
\arguments{
\item{survey1}{index column of first data set}

\item{survey2}{index column of second data set}

\item{year}{optional column of years if you want to split data by years}
}
\description{
Normalizer function to put two different data sets on the same scale in a ridgeline plot.
}
