% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{match_normalizer2}
\alias{match_normalizer2}
\title{Normalizer function to put multiple different data sets on the same scale in a ridgeline plot}
\usage{
match_normalizer2(..., year = NULL)
}
\arguments{
\item{...}{index columns of the data sets}

\item{year}{optional column of years if you want to split data by years}
}
\description{
Normalizer function to put multiple different data sets on the same scale in a ridgeline plot
}
