% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{match_TMB_names}
\alias{match_TMB_names}
\title{Match TMB names with LaTeX or TeX names}
\usage{
match_TMB_names(
  var_names,
  fixed = rep(FALSE, length(var_names)),
  type,
  transformed
)
}
\arguments{
\item{var_names}{vector of names to match}

\item{fixed}{boolean vector indicating whether or not a value is fixed}

\item{type}{either LaTeX or TeX}

\item{transformed}{use transformed form or not}
}
\description{
Match TMB names with LaTeX or TeX names
}
