% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{ordered_inv_transform}
\alias{ordered_inv_transform}
\title{Transform vector back to monotone increasing}
\usage{
ordered_inv_transform(y)
}
\arguments{
\item{y}{the vector to transform back}
}
\description{
Transform vector back to monotone increasing
}
