% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{ordered_transform}
\alias{ordered_transform}
\title{Transform monotone increasing vector to optimizer scale}
\usage{
ordered_transform(x)
}
\arguments{
\item{x}{the vector to transform}
}
\description{
Transform monotone increasing vector to optimizer scale
}
