% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{plotly_bubbles}
\alias{plotly_bubbles}
\title{Create a bubble plot using plotly}
\usage{
plotly_bubbles(data, x, y, bub_z, title, xlab, ylab)
}
\arguments{
\item{data}{the data to plot}

\item{x}{the x variable column}

\item{y}{the y variable column}

\item{bub_z}{the variable to make into a bubble!}

\item{title}{the title of the plot}

\item{xlab}{the name of the x axis}

\item{ylab}{the name of the y axis}
}
\description{
Create a bubble plot using plotly
}
