% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{plotly_corr_mat}
\alias{plotly_corr_mat}
\title{plotly correlation matrix}
\usage{
plotly_corr_mat(cov.matrix, title, cov.names = NULL)
}
\arguments{
\item{cov.matrix}{the covariance matrix to plot as correlation matrix}

\item{title}{Title of the plot}

\item{cov.names}{the names you want for cov.matrix (optional)}
}
\description{
plotly correlation matrix
}
