% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{plotly_pla}
\alias{plotly_pla}
\title{Plot Probability Length at Age Matrix}
\usage{
plotly_pla(plamat, ages, lengths, len_mus, len_sds, title, crit_value = 1.96)
}
\arguments{
\item{plamat}{the PLA at age matrix to plot}

\item{ages}{the vector of ages in the model}

\item{lengths}{the vector of lengths to use in the model}

\item{len_mus}{the vector of predicted mean lengths at age}

\item{len_sds}{the vector of predicted sd lengths at age}

\item{title}{the title for the plot}

\item{crit_value}{the critical value to use}
}
\description{
Plot Probability Length at Age Matrix
}
