% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{plotly_resids}
\alias{plotly_resids}
\title{plotly residuals}
\usage{
plotly_resids(data, x, y, title, xlab, ylab)
}
\arguments{
\item{data}{the data to be plotted}

\item{x}{the x variable}

\item{y}{the y variable}

\item{title}{the overall title}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}
}
\description{
Plot residuals with horizontal line at zero and trendline
}
