% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfplots.R
\name{plotly_ridges}
\alias{plotly_ridges}
\title{Ridgeline Plot using Plotly}
\usage{
plotly_ridges(
  data,
  x,
  y,
  height,
  xlab,
  ylab,
  heightlab,
  main,
  plot_group,
  plot_color,
  normalizer = function(x) {
     (x - min(x))/(max(x) - min(x))
 },
  n_group = NULL,
  add = NULL
)
}
\arguments{
\item{data}{the dataframe to plot}

\item{x}{the x variable (e.g., length,age)}

\item{y}{the y variable (e.g., year)}

\item{height}{the ridgeline height variable (e.g., survey index)}

\item{xlab}{X axis label name (also used in hovertext)}

\item{ylab}{Y axis label name}

\item{heightlab}{name of the height variable (used in hovertext)}

\item{main}{main plot title}

\item{plot_group}{label for legend if plotting multiple}

\item{plot_color}{color of plot}

\item{normalizer}{the function to do the normalizing}

\item{n_group}{how to do the grouping for the normalizer}

\item{add}{prior plotly object if adding on to existing plot}
}
\description{
Ridgeline plot using plotly. This allows combining different plots together and adjusting the scale via
a user specified normalizer function.
}
