% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{proj_report}
\alias{proj_report}
\title{Projection Report}
\usage{
proj_report(proj.object, report_output, title = "ProjReport")
}
\arguments{
\item{proj.object}{the object containing the projections}

\item{title}{name of the file to make}

\item{report_ouput}{the path to put the report}
}
\description{
Create a report for a set of projections
}
