% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{redfish_rat}
\alias{redfish_rat}
\title{Generate prior for conv. factor}
\usage{
redfish_rat(x)
}
\arguments{
\item{x}{the value of the length}
}
\description{
This uses the inverse of the conversion factor between
Engel and Campelen Trawl from Warren 1997
}
