% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{retro_report}
\alias{retro_report}
\title{Retro Report}
\usage{
retro_report(peels.object, report_output, title = "RetroReport")
}
\arguments{
\item{peels.object}{the object containing the peels}

\item{report_output}{the path to put the report}

\item{title}{name of the file to make}
}
\description{
Create a report for the retros run, plots SSB, Average F, survey Abundance, abundance at age
}
