% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{run_peels}
\alias{run_peels}
\title{Run retrospective peels}
\usage{
run_peels(
  model.object,
  n_peels = 6,
  use_sdr = TRUE,
  quick_start = TRUE,
  silence = TRUE
)
}
\arguments{
\item{model.object}{the output of run_redfish_acl}

\item{n_peels}{the number of years to peel back, including the original}

\item{use_sdr}{use sdreport or not?}

\item{quick_start}{use fixed effect parameters from original model fit as starting values or use defaults (false)}

\item{silence}{disable TMB tracing for each peel?}
}
\description{
Construct models removing the number of specified years. Returns a list of model objects for each of the last years removed.
}
