% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{run_projections}
\alias{run_projections}
\title{Run Projections on the model}
\usage{
run_projections(
  model.object,
  proj_years = 3,
  weights,
  maturity,
  supplied_F = NULL,
  given_catch = NULL
)
}
\arguments{
\item{model.object}{the output of run_redfish_acl}

\item{proj_years}{number of years to project}

\item{weights}{the weights at length for the projection years}

\item{maturity}{the maturity ogives at length for the projection years}

\item{supplied_F}{optional supplied vector to use for log F_y}

\item{given_catch}{optional vector of catch to use for projections}
}
\description{
Project the model based on certain assumptions. If both supplied_F and given_catch are null we
assume to just use the terminal year. Projections are currently only implemented for models using
'simple_blocked_len' selectivity and those with cohort effects.
}
