% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{run_redfish_acl}
\alias{run_redfish_acl}
\title{Run Model}
\usage{
run_redfish_acl(
  weight_array,
  maturity_array,
  survey_df,
  landings_df,
  F_type,
  N_type,
  M_type,
  catch_type,
  landing_type,
  base_M,
  output_path = NULL,
  report_title = NULL,
  catch_prop = NULL,
  agg_key = NULL,
  upper_landing_bounds = NULL,
  lower_landing_bounds = NULL,
  years = 1983:2021,
  ages = 1:20,
  lengths = 7:45,
  tmb.map = NULL,
  random = NULL,
  control = NULL,
  start.parms = NULL,
  data = NULL,
  model_num = NULL,
  catch_sig_type = "independent",
  survey_sig_type = "independent",
  inf_length = 60,
  optimize = TRUE,
  use_sdreport = TRUE,
  plot_output = FALSE,
  use_Qpriors = FALSE,
  fix_L_inf = NULL,
  rand_N = TRUE,
  Q_prior_max = 35,
  inter_hessian = FALSE,
  pg_ext = 200,
  gf_type = "avgage",
  inner.control = list(smartsearch = TRUE),
  silent = FALSE,
  rounding_bit = 0.075,
  double_opt = FALSE,
  dopt_control = NULL,
  survey_sd_map = NULL,
  catch_prop_map = NULL
)
}
\arguments{
\item{weight_array}{array of weights with dimnames for male/female, year, length, etc.}

\item{maturity_array}{array of maturity ogives to use, with appropriate dimnames}

\item{survey_df}{data.frame of survey data to use}

\item{landings_df}{data.frame of landings data}

\item{F_type}{either 'age' or 'simple_blocked_len'}

\item{N_type}{either 'fixed' or 'cohort'}

\item{M_type}{Either 'fixed', 'M-devs', 'Lorenzen' or 'LorenzenDevs'}

\item{catch_type}{Either 'no_prop' or 'prop'}

\item{landing_type}{Either 'unbounded' or 'bounded'}

\item{base_M}{the base_M underlying value assumption}

\item{output_path}{path to put the report and model output/data (optional)}

\item{report_title}{what to call the report (optional)}

\item{catch_prop}{optional matrix with named row/cols of catch proportions}

\item{agg_key}{required if using catch_prop, specifies how to aggregate catch}

\item{upper_landing_bounds}{column of landings upper multiplier for bounds (only for bounded landings)}

\item{lower_landing_bounds}{column of landings lower multiplier for bounds (only for bounded landings)}

\item{years}{vector of years to include}

\item{ages}{vector of ages to use in the model}

\item{lengths}{vector of lengths to include}

\item{tmb.map}{optional named list, use to manually provide TMB map}

\item{random}{optional, override the default random effect parameters}

\item{control}{optional, override nlminb optimizer settings (see nlminb help)}

\item{start.parms}{optional named list of start parameters, used to override the starting parameters}

\item{data}{optional named list of data to give the TMB model, used to directly override the data}

\item{model_num}{optional model number for identity purposes}

\item{catch_sig_type}{Either 'independent', 'AR1' or 'compound'}

\item{survey_sig_type}{Either 'independent' or 'AR1'}

\item{inf_length}{the max length used in the model}

\item{optimize}{optimize the model?}

\item{use_sdreport}{run the sdreport?}

\item{plot_output}{generate the plot html?}

\item{use_Qpriors}{use the data based prior on survey Q}

\item{fix_L_inf}{if given a value in cm this will be used as the value of L_inf, setting L to the corresponding value}

\item{rand_N}{run the random N draw?}

\item{Q_prior_max}{the maximum of the Q prior length to use}

\item{inter_hessian}{use internal hessian instead of optimized}

\item{pg_ext}{controls how fine the plus group growth extension goes}

\item{gf_type}{controls some settings on the growth function 'noavgage' has no plus group extension, 'avgage' does.}

\item{inner.control}{inner.control to pass to MakeADFun}

\item{silent}{trace the model or not}

\item{rounding_bit}{how much 95\% selectivity for catch is ahead of 50\%}

\item{double_opt}{run the optmizer again for a bit?}

\item{dopt_control}{control for the second round of optimiztion}

\item{survey_sd_map}{list by year and length specifying how to internally map the survey sds}

\item{catch_prop_map}{list by year and length specifying how to internally map the catch prop sds}
}
\description{
Run the model with the selected settings
}
