% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{simplex_transform}
\alias{simplex_transform}
\title{Simplex transform with any maximum}
\usage{
simplex_transform(x, L)
}
\arguments{
\item{x}{the vector to transform}

\item{L}{the maximum to hit}
}
\description{
Simplex transform with any maximum
}
