% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{unit_simplex_transform}
\alias{unit_simplex_transform}
\title{Convert a vector that sums to 1 to be on all real space}
\usage{
unit_simplex_transform(x)
}
\arguments{
\item{x}{the vector of porportions}
}
\description{
Convert a vector that sums to 1 to be on all real space
}
