% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfhelps.R
\name{untransform_names}
\alias{untransform_names}
\title{Convert to untransformed names}
\usage{
untransform_names(var_names)
}
\arguments{
\item{var_names}{vector of variable names}
}
\description{
Removes logs and ts and whatever from TMB names
}
